+++ src/search/mod.rs
@@ -1,3 +1,5 @@
+pub mod metasearch_api;
 pub mod plan;
 
+pub use metasearch_api::*;
 pub use plan::*;